+++ lib/linea-verifier/Cargo.toml
@@ -12,17 +12,18 @@ workspace = true
 test-include = ["lib/scroll-verifier/tests"]
 
 [dependencies]
-ethereum-verifier = { workspace = true }
-ethers-core       = { workspace = true }
-gnark-mimc        = { workspace = true }
-hex               = { workspace = true }
-hex-literal       = { workspace = true }
-linea-zktrie      = { workspace = true }
-rlp               = { workspace = true }
-scroll-codec      = { workspace = true }
-serde             = { workspace = true }
-serde-utils       = { workspace = true }
-serde_json        = { workspace = true }
-sha3              = { workspace = true }
-thiserror         = { workspace = true }
-unionlabs         = { workspace = true }
+ethereum-verifier        = { workspace = true }
+ethers-core              = { workspace = true }
+gnark-mimc               = { workspace = true }
+hex                      = { workspace = true }
+hex-literal              = { workspace = true }
+linea-light-client-types = { workspace = true }
+linea-zktrie             = { workspace = true }
+rlp                      = { workspace = true }
+scroll-codec             = { workspace = true }
+serde                    = { workspace = true }
+serde-utils              = { workspace = true }
+serde_json               = { workspace = true }
+sha3                     = { workspace = true }
+thiserror                = { workspace = true }
+unionlabs                = { workspace = true }
