+++ lib/unionlabs/src/ibc/lightclients.rs
@@ -1,6 +1,5 @@
 pub mod ethereum;
 pub mod evm_in_cosmos;
-pub mod linea;
 pub mod movement;
 pub mod scroll;
 pub mod wasm;
