+++ light-clients/linea-light-client/Cargo.toml
@@ -13,24 +13,25 @@ workspace = true
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
-cosmwasm-std          = { workspace = true, features = ["abort"] }
-ethereum-light-client = { workspace = true, features = ["mainnet", "library"] }
-ethereum-verifier     = { workspace = true }
-ethers-core           = { workspace = true }
-gnark-mimc            = { workspace = true }
-hex                   = { workspace = true }
-ics008-wasm-client    = { workspace = true }
-linea-verifier        = { workspace = true }
-linea-zktrie          = { workspace = true }
-protos                = { workspace = true }
-rlp                   = { workspace = true }
-schemars              = { workspace = true }
-serde                 = { workspace = true, features = ["derive"] }
-serde-json-wasm       = { workspace = true }
-sha3                  = { workspace = true }
-thiserror             = { workspace = true }
-tiny-keccak           = { workspace = true, features = ["keccak"] }
-unionlabs             = { workspace = true, features = ["ethabi", "stargate"] }
+cosmwasm-std             = { workspace = true, features = ["abort"] }
+ethereum-light-client    = { workspace = true, features = ["mainnet", "library"] }
+ethereum-verifier        = { workspace = true }
+ethers-core              = { workspace = true }
+gnark-mimc               = { workspace = true }
+hex                      = { workspace = true }
+ics008-wasm-client       = { workspace = true }
+linea-light-client-types = { workspace = true, features = ["proto"] }
+linea-verifier           = { workspace = true }
+linea-zktrie             = { workspace = true }
+protos                   = { workspace = true }
+rlp                      = { workspace = true }
+schemars                 = { workspace = true }
+serde                    = { workspace = true, features = ["derive"] }
+serde-json-wasm          = { workspace = true }
+sha3                     = { workspace = true }
+thiserror                = { workspace = true }
+tiny-keccak              = { workspace = true, features = ["keccak"] }
+unionlabs                = { workspace = true, features = ["ethabi", "stargate"] }
 
 [dev-dependencies]
 base64     = { workspace = true }
