+++ light-clients/linea-light-client/types/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+edition = "2021"
+name    = "linea-light-client-types"
+version = "0.1.0"
+
+[dependencies]
+protos    = { workspace = true, features = ["union+ibc+lightclients+linea+v1"] }
+serde     = { workspace = true, features = ["derive"] }
+thiserror = { workspace = true }
+unionlabs = { workspace = true }
+
+[features]
+default = ["proto"]
+proto   = ["unionlabs/proto"]
+
+[dev-dependencies]
+hex-literal = { workspace = true }
